+++ mi-ui/src/components/PageComponents/Home/LoginScreen/index.tsx
@@ -1,11 +1,11 @@
 import DarkModeToggle from "@components/Layout/Header/DarkModeToggle";
+import News from "@components/SharedComponents/News";
 import { LeaderboardType, NewsType } from "@libs/types/influence";
 import { FC } from "react";
 import { useGlobalTooltip } from "src/states/globalTooltip";
 
-import ContributeButtons from "../Shared/ContributeButtons";
-import Leaderboard from "../Shared/Leaderboard";
-import News from "../Shared/News";
+import ContributeButtons from "../../../SharedComponents/ContributeButtons";
+import Leaderboard from "../../../SharedComponents/Leaderboard";
 import CoolCards from "./CoolCards";
 import styles from "./style.module.scss";
 
@@ -42,8 +42,7 @@ const LoginScreen: FC<Props> = ({ topList, newsList }) => {
             className={styles.a}
             onMouseEnter={(e) =>
               activateTooltip("Opens in new tab", e.currentTarget)
-            }
-          >
+            }>
             pishifat’s Mapper Influences
           </a>{" "}
           project.
