+++ mi-ui/src/components/SharedComponents/Leaderboard/style.module.scss
@@ -19,7 +19,7 @@
     padding-bottom: 0;
     width: 100%;
     max-height: 22rem;
-    overflow-y: scroll;
+    overflow-y: auto;
 
     .row {
       display: flex;
