+++ mi-ui/src/components/SharedComponents/News/NewsRow/style.module.scss
